feat: MCP performance optimizations and middleware logging fix
- Add context support for MCP tools with 30s timeout
- Implement resource caching with 5-minute TTL
- Add comprehensive metrics collection for MCP operations
- Fix middleware registration logging spam
- Update documentation for STDIO transport
- Fix calculator tool to handle infinity/NaN results
- Add integration tests for MCP optimizations

+++ examples/configuration/main.go
@@ -206,7 +206,7 @@ func runCombinedConfig() {
 	fmt.Println()
 	fmt.Println("Final configuration (after precedence):")
 	fmt.Printf("  Port: %s (from environment)\n", server.Options.Addr)
-	fmt.Printf("  Rate Limit: %d req/s (from JSON)\n", server.Options.RateLimit)
+	fmt.Printf("  Rate Limit: %.0f req/s (from JSON)\n", float64(server.Options.RateLimit))
 	fmt.Printf("  Burst Limit: %d (from JSON)\n", server.Options.Burst)
 
 	// Cleanup
